requirements.txt
sqlparse
pandas
sql_metadata
openai
cx-Oracle
mysql-connector-python